# internal/http/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API for the Codevertex inventory service. Provides catalog, warehouse, and inventory management capabilities.",
        "title": "Inventory Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/{tenant}/inventory/stock/export": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Stock levels export document (PDF/CSV)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pdf | csv",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "item/SKU/warehouse search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by category",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "GOODS | INGREDIENT | EQUIPMENT",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter to one warehouse",
                        "name": "warehouse_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter to one warehouse sub-location (zone/aisle/rack/shelf/bin)",
                        "name": "location_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "override the ambient X-Outlet-ID scope",
                        "name": "outlet_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "only items at/below reorder level",
                        "name": "low_stock",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "only items with zero available",
                        "name": "out_of_stock",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "category — render one section per category instead of one flat table",
                        "name": "group_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearerAuth": {
            "description": "JWT token from auth-service. Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}